requirements.txt
numpy==1.21.2
Flask==2.0.1
logger==1.4
pandas==1.3.2
xgboost==1.5.0
Flask_Cors==3.0.10
cassandra_driver==3.25.0
imblearn==0.0
scikit_learn==1.0.1
gunicorn